+++ lib/grape/middleware/base.rb
@@ -11,7 +11,7 @@ module Grape
       # @param [Hash] options A hash of options, simply stored for use by subclasses.
       def initialize(app, **options)
         @app = app
-        @options = merge_default_options(options)
+        @options = merge_default_options(options).freeze
         @app_response = nil
       end